refactor: reorganize fs structure for android
