+++ android/src/fabric/java/com/reactnativekeyboardcontroller/KeyboardControllerModule.kt
@@ -1,6 +1,7 @@
 package com.reactnativekeyboardcontroller
 
 import com.facebook.react.bridge.ReactApplicationContext
+import com.reactnativekeyboardcontroller.modules.KeyboardControllerModuleImpl
 
 class KeyboardControllerModule(mReactContext: ReactApplicationContext) : NativeKeyboardControllerSpec(mReactContext) {
   private val module = KeyboardControllerModuleImpl(mReactContext)
