+++ android/src/main/java/com/reactnativekeyboardcontroller/extensions/Float.kt
@@ -1,0 +1,8 @@
+package com.reactnativekeyboardcontroller.extensions
+
+import android.content.res.Resources
+
+val Float.dp: Int
+  get() = (this / Resources.getSystem().displayMetrics.density).toInt()
+val Float.px: Int
+  get() = (this * Resources.getSystem().displayMetrics.density).toInt()
