+++ android/src/main/java/com/reactnativekeyboardcontroller/modules/StatusBarManagerCompatModuleImpl.kt
@@ -1,4 +1,4 @@
-package com.reactnativekeyboardcontroller
+package com.reactnativekeyboardcontroller.modules
 
 import android.animation.ArgbEvaluator
 import android.animation.ValueAnimator
@@ -10,8 +10,8 @@ import androidx.core.view.WindowInsetsControllerCompat
 import com.facebook.react.bridge.ReactApplicationContext
 import com.facebook.react.bridge.UiThreadUtil
 
-class StatusBarManagerCompatImpl(private val mReactContext: ReactApplicationContext) {
-  private val TAG = StatusBarManagerCompatImpl::class.qualifiedName
+class StatusBarManagerCompatModuleImpl(private val mReactContext: ReactApplicationContext) {
+  private val TAG = StatusBarManagerCompatModuleImpl::class.qualifiedName
   private var controller: WindowInsetsControllerCompat? = null
 
   fun setHidden(hidden: Boolean) {
